separate worker, clean up code
