+++ app/src/main/java/com/example/firstapp/SplashActivity.kt
@@ -6,15 +6,16 @@ import android.os.Handler
 import android.os.Looper
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
+import androidx.work.OneTimeWorkRequest
 import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkManager
 import androidx.work.WorkRequest
 import com.example.firstapp.databinding.ActivitySplashBinding
 import com.example.firstapp.eventbus.EventBus
 import com.example.firstapp.eventbus.InitDataEvent
-import com.example.firstapp.worker.InitDataWorker
+import com.example.firstapp.worker.InitChampionWorker
+import com.example.firstapp.worker.InitItemWorker
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
 import timber.log.Timber
@@ -53,8 +54,12 @@ class SplashActivity : AppCompatActivity() {
             EventBus.subscribe<InitDataEvent>().collectLatest { updateProgress(it) }
         }
 
-        val initDataWorkRequest: WorkRequest = OneTimeWorkRequestBuilder<InitDataWorker>().build()
-        WorkManager.getInstance(applicationContext).enqueue(initDataWorkRequest)
+        val initChampionWorkRequest: WorkRequest = OneTimeWorkRequestBuilder<InitChampionWorker>().build()
+        val initItemWorkRequest: WorkRequest = OneTimeWorkRequestBuilder<InitItemWorker>().build()
+        WorkManager.getInstance(applicationContext)
+                .beginWith(initChampionWorkRequest as OneTimeWorkRequest)
+                .then(initItemWorkRequest as OneTimeWorkRequest)
+                .enqueue()
     }
 
 }
