+++ app/src/main/res/values/strings.xml
@@ -72,4 +72,6 @@
 
     <string name="item_name">아이템 이름</string>
     <string name="finish">완료</string>
+    <string name="init_champion_start">챔피온 정보를 초기화합니다.</string>
+    <string name="init_item_start">아이템 정보를 초기화합니다.</string>
 </resources>
